Update MFA_NPS_Troubleshooter.ps1
There appears to be a bug in this registry check. It currently checks to see if AZURE_MFA_HOSTNAME = adnotifications.windowsazure.com, which will always cause it to fail. It should either confirm that AZURE_MFA_HOSTNAME = strongauthenticationservice.auth.microsoft.com and/or that AZURE_MFA_RESOURCE_HOSTNAME = adnotifications.windowsazure.com.